requirements.txt
adhanpy>=1.0.0
APScheduler>=3.10,<4
requests>=2.28
PyYAML>=6.0
